# docs/api-reference/schemas/openapi-clob.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CLOB API",
    "version": "1.0.0",
    "description": "Specification for the central limit order book service. Import this file into Paw (RapidAPI Client) to explore and test the available endpoints."
  },
  "servers": [
    {
      "url": "https://clob.kuest.com",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/data/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "List live orders for the authenticated account",
        "description": "Returns open orders for the authenticated account with optional market and asset filters.",
        "operationId": "listOpenOrders",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "description": "Filter by order identifier.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "market",
            "in": "query",
            "required": false,
            "description": "Filter by market/condition identifier.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "asset_id",
            "in": "query",
            "required": false,
            "description": "Filter by asset (token) identifier.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "next_cursor",
            "in": "query",
            "required": false,
            "description": "Base64-encoded offset for pagination. Use `MA==` to start; an empty response cursor means no more pages.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Open orders owned by the authenticated account.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "next_cursor", "limit", "count"],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Order"
                      }
                    },
                    "next_cursor": {
                      "type": "string"
                    },
                    "limit": {
                      "type": "integer",
                      "description": "Page size used for the query.",
                      "example": 100
                    },
                    "count": {
                      "type": "integer",
                      "description": "Number of items returned in this page."
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
...
    "responses": {
      "ErrorResponse": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message."
          }
        }
      },
...
      "Order": {
        "type": "object",
        "required": [
          "associateTrades",
          "id",
          "status",
          "market",
          "originalSize",
          "outcome",
          "makerAddress",
          "side",
          "sizeMatched",
          "assetId",
          "type",
          "createdAt"
        ],
        "properties": {
          "associateTrades": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Identifiers of trades that include this order."
          },
          "id": {
            "type": "string",
            "description": "Order ULID assigned by the engine."
          },
          "status": {
            "type": "string",
            "description": "Current order status (`live`, `matched`, `delayed`, `unmatched`)."
          },
          "market": {
            "type": "string",
            "description": "Market/condition identifier."
          },
          "originalSize": {
            "type": "string",
            "description": "Original submitted size, in whole shares."
          },
          "outcome": {
            "type": "string",
            "description": "Outcome label (token identifier)."
          },
          "makerAddress": {
            "type": "string",
            "description": "Address providing maker liquidity."
          },
          "owner": {
            "type": "string",
            "nullable": true,
            "description": "API key owner when recorded."
          },
          "price": {
            "type": "string",
            "nullable": true,
            "description": "Limit price (null when indeterminate)."
          },
          "side": {
            "type": "string",
            "enum": ["BUY", "SELL"],
            "description": "Human-readable side of the order book."
          },
          "sizeMatched": {
            "type": "string",
            "description": "Total matched size so far, in whole shares."
          },
          "assetId": {
            "type": "string",
            "description": "Token identifier."
          },
          "type": {
            "type": "string",
            "enum": ["FOK", "FAK", "GTC", "GTD"],
            "description": "Time-in-force policy recorded for the order."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}